Modularize segy ingestion code
