+++ src/mdio/ingestion/segy/__init__.py
@@ -1,0 +1,1 @@
+"""SEG-Y specific ingestion helpers."""
